moved SingletonModuleRegistry to test package
